Move `DeviceListResyncTestCase` to their own file
